refactor: isolate device identity repair
